Refactor product factories and tests; consolidate fixtures
- Created a new factories file for product-related tests to improve organization and maintainability.
- Removed the old product factories file and migrated relevant functions to the new factories file.
- Introduced a new fixtures file for products to centralize test data and reduce duplication.
- Updated all test files to reference the new fixtures and factories structure.
- Adjusted solution list page and tutorial tests to align with the new data structure.
- Enhanced webinar tests with new factory functions for handling edge cases.
