+++ apps/nextjs-website/src/lib/guideListPages/__tests__/fixtures.ts
@@ -1,6 +1,6 @@
 import { StrapiGuideListPages } from '@/lib/guideListPages/types';
 import { GuideListPageProps } from '@/app/[locale]/[productSlug]/guides/page';
-import { product } from '@/lib/products/__tests__/fixtures/product';
+import { product } from '@/lib/products/__tests__/fixtures';
 
 const dateNow = new Date().toISOString();
