+++ apps/nextjs-website/src/lib/solutionListPage/__tests__/factories.ts
@@ -1,74 +1,75 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { strapiSolutionListPage } from '@/lib/solutionListPage/__tests__/fixtures';
 import type { StrapiSolutionListPage } from '../types';
+import { wrapAsRootEntity } from '../../__tests__/strapiEntityWrappers';
 
 export function minimalDataSolutionListPage(): StrapiSolutionListPage {
-  return {
+  return wrapAsRootEntity({
     title: 'Minimal Solutions',
     description: 'Minimal solutions page',
     solutions: [],
     caseHistories: undefined,
     features: undefined,
     seo: undefined,
-  };
+  });
 }
 
 export function solutionListPageWithoutCaseHistories(): StrapiSolutionListPage {
-  return {
-    ...strapiSolutionListPage,
+  return wrapAsRootEntity({
+    ...strapiSolutionListPage.data,
     caseHistories: undefined,
-  };
+  });
 }
 
 export function solutionListPageWithoutFeatures(): StrapiSolutionListPage {
-  return {
-    ...strapiSolutionListPage,
+  return wrapAsRootEntity({
+    ...strapiSolutionListPage.data,
     features: undefined,
-  };
+  });
 }
 
 export function solutionListPageWithoutSolutions(): StrapiSolutionListPage {
-  return {
-    ...strapiSolutionListPage,
+  return wrapAsRootEntity({
+    ...strapiSolutionListPage.data,
     solutions: [],
-  };
+  });
 }
 
 export function solutionListPageWithMissingSolutionSlug(): StrapiSolutionListPage {
-  return {
-    ...strapiSolutionListPage,
+  return wrapAsRootEntity({
+    ...strapiSolutionListPage.data,
     solutions: [
       {
-        ...strapiSolutionListPage.solutions[0],
+        ...strapiSolutionListPage.data.solutions[0],
         slug: undefined as unknown as string,
         title: 'Solution Without Slug',
       },
       {
-        ...strapiSolutionListPage.solutions[0],
+        ...strapiSolutionListPage.data.solutions[0],
         title: 'Valid Solution',
       },
     ],
-  };
+  });
 }
 
 export function solutionListPageWithMissingCaseHistorySlug(): StrapiSolutionListPage {
-  return {
-    ...strapiSolutionListPage,
+  return wrapAsRootEntity({
+    ...strapiSolutionListPage.data,
     caseHistories: {
-      ...strapiSolutionListPage.caseHistories,
+      ...strapiSolutionListPage.data.caseHistories,
       case_histories: [
         {
-          ...strapiSolutionListPage.caseHistories?.case_histories[0],
+          ...strapiSolutionListPage.data.caseHistories?.case_histories[0],
           id: 1,
           slug: undefined as unknown as string,
           title: 'Case History Without Slug',
         },
         {
-          ...strapiSolutionListPage.caseHistories?.case_histories[0],
+          ...strapiSolutionListPage.data.caseHistories?.case_histories[0],
           id: 2,
           title: 'Valid Case History',
         },
       ],
     },
-  };
+  });
 }
