+++ apps/nextjs-website/src/lib/solutionListPage/mapper.ts
@@ -1,12 +1,11 @@
 /* eslint-disable functional/no-expression-statements */
 import { SolutionListTemplateProps } from '@/components/templates/SolutionListTemplate/SolutionListTemplate';
-import { RootEntity } from '@/lib/strapi/types/rootEntity';
 import { compact } from 'lodash';
 import { StrapiSolutionListPage } from './types';
 
 export function mapSolutionListPageProps(
   locale: string,
-  strapiSolutionsList: RootEntity<StrapiSolutionListPage>
+  strapiSolutionsList: StrapiSolutionListPage
 ): SolutionListTemplateProps {
   const strapiSolutionsListData = strapiSolutionsList.data;
