+++ apps/nextjs-website/src/lib/solutionListPage/types.ts
@@ -2,12 +2,13 @@ import { CaseHistoriesComponent } from '@/lib/caseHistories/types';
 import { StrapiBaseSolutionWithProducts } from '@/lib/solutions/types';
 import { StrapiFeatures } from '@/lib/strapi/types/features';
 import type { StrapiSeo } from '@/lib/seo/strapiTypes';
+import { RootEntity } from '@/lib/strapi/types/rootEntity';
 
-export type StrapiSolutionListPage = {
+export type StrapiSolutionListPage = RootEntity<{
   readonly title: string;
   readonly description: string;
   readonly caseHistories?: CaseHistoriesComponent;
   readonly solutions: readonly StrapiBaseSolutionWithProducts[];
   readonly features?: StrapiFeatures;
   readonly seo?: StrapiSeo;
-};
+}>;
